SERVER-100441 Prevents aborting a txn when no txn was started (#32161)
In legacy WouldChangeOwningShardError handler for mongos findAndModify
command, where the command is a retryable write (not in a user-started
transaction), the exception handler assumes that the findAndModify
internal transaction was started and therefore tries to abort the
transaction. However, an exception may be thrown before the internal
transaction was started; in this case, the exception handler must not
attempt to abort the (non-existent) transaction.