SERVER-82025 No-op updates should complain about unsupported "c" field for non-pipeline updates (#33389)
Each update statement inside an `update` command can have a `c` field
with constant values.
[Our
documentation](https://www.mongodb.com/docs/manual/reference/command/update/#std-label-update-command-c)
states the following about the `c` field:
> Optional. You can specify c only if
[u](https://www.mongodb.com/docs/manual/reference/command/update/#std-label-update-command-u)
is a pipeline.

This constraint however was not enforced for no-op updates, e.g. when
the document to update did not exist.
This PR will enforce the constraint for all update operations,
regardless of whether they are no-ops or actual updates.

In case the `c` field is specified and the `u` field does not contain a
pipeline update, error code `51198` with message `Constant values may
only be specified for pipeline updates` is now consistently returned.
Previously it was only returned for updates that were actually executed.