SERVER-91164 Expand property-based testing $group and wildcard index models (#33390)
$group is modeled more cleanly. It can generate zero to three
accumulators, rather than exactly one every time. This means we're more
likely to run into distinct scan plans.

The index models are cleaned up a bit, and additional cases for wildcard
indexes are supported. Before, the only type of wildcard index supported
was `{"$**": 1}`. Now we have compound WC indexes, as well as WC indexes
on a dotted path, such as `{a: 1, "b.$**": -1, c: 1}`

This catches
[SERVER-102075](https://jira.mongodb.org/browse/SERVER-102075) and
[SERVER-102229](https://jira.mongodb.org/browse/SERVER-102229)