SERVER-95374 Strip invalid compound wildcard index assignment in rated MatchExpression tree (#32600)
This commit addresses a bug related to handling invalid compound
wildcard index (CWI) assignments within query trees. A CWI assignment is
deemed invalid if it does not involve a non-$_path expanded wildcard
field. Such assignments are unnecessary because their index bounds are
equivalent to those derived from $_path expanded CWIs. Additionally,
they complicate the construction of index bounds and require further
correction that were handled by `expandWildcardFieldBounds()`. However,
in some cases, `expandWildcardFieldBounds()` may incorrectly overwrite
the index bounds and lead to incorrect query results.

To address the bug, this commit introduces early stripping of invalid
CWI assignments within `stripInvalidAssignments()`. This not only
prevents enumerating invalid query plans but also prevents the need for
post-processing corrections on the index bounds. The function
`stripInvalidAssignmentsToCompoundWildcardIndexes()` is introduced to
search CWI assignments and strip them if they are invalid. Further, it
takes the conjunctive nodes into account, ensuring that the AND-related
predicates are evaluated collectively and considered valid if any
predicate generates bounds on the wildcard field.

Following this change, `expandWildcardFieldBounds()` is removed, as it
is no longer necessary due to the early stripping of invalid CWI
assignments.

---------

Co-authored-by: David Storch <dstorch@users.noreply.github.com>