SERVER-93194: Add diagnostic printer for QueryPlannerParams (#33380)
Before query planning, a 'ScopedDebugInfo' captures a reference to the
'QueryPlannerParams', allowing us to log the indexes considered during
planning if we encounter a tassert/invariant/segfault.

This commit includes the following changes:

- A new diagnostic printer, 'QueryPlannerParamsPrinter', is introduced.
- A 'ScopedDebugInfo' for the 'QueryPlannerParamsPrinter' is built after
a 'PrepareExecutionHelper' is constructed and prior to planning. It has
support for serializing just the owned values of 'IndexEntry' .