SERVER-101752 Unit test buildBatchedInsertContexts, stageInsertBatch, stageInsertBatchIntoEligibleBuckets for measurements without meta values to a collection with a meta field (#33310)
also covers
[SERVER-101634](https://jira.mongodb.org/browse/SERVER-101634) (will
close as gone away)

- For `stageInsertBatch` and `stageInsertBatchIntoEligibleBucket`, when
we have no meta fields in measurements, adds coverage by testing in both
collections with and without meta fields.
- Changes some tests for `stageInsertBatch` and
`stageInsertBatchIntoEligibleBucket` to expand coverage of checking no
meta fields measurements in collections without meta fields.
- Split the RolloverMixed (both for `stageInsertBatch` and
`stageInsertBatchIntoEligibleBucket` into three different sets of tests:
   1. With Meta (has to be collection with a meta field)
   2. Without Meta (tests both collection with and without meta field)
3. With Cache Pressure (this unit test purposely lower the
`kStorageCacheSizeBytes` to initiate the `kCachePressure` rollover
reason, so I thought it would make sense to distinguish this)
- For the tests I touched with `fromjson()` declared measurements,
change them to `BSON()`
- Removed erasing from `expectedIndicesWithErrors` in
`_testBuildBatchedInsertContextWithoutMetaField` and
`_testBuildBatchedInsertContextWithMetaField` because we check the size
is the same and the user of this function should be using distinct
indices. I added this as a comment

---------

Co-authored-by: Matt Kneiser <matt.kneiser@mongodb.com>