SERVER-102319 Fix hard-closed rollover actions stats being incorrect in new timeseries write path (#33623)
- Moved `updateRolloverStats` to happen whenever we are in the
`allCommitted` branch of rollover.
- In the previous timeseries write path, we could have some stats
updated more eagerly if there is a bucket without all batches committed
but we have called rollover, because we updated the stats in
determineRolloverReason.
- After this change, we update the hard closed rollover stats only when
we know that we have rolled-over a bucket with all committed all of our
measurements in the bucket (if we are attempting to rollover a bucket
that doesn't have all measurements committed, we just set the
rolloverReason of the bucket without any further action).
- Added unit testing -- made it so that we had to move the helper
function `_getExecutionStat` from `BucketCatalogTest` -->
`TimeseriesTestFixture`
- Made `BucketCatalogInternalTest` inherit from `TimeseriesTestFixture`
- Created [follow-up
ticket](https://jira.mongodb.org/browse/SERVER-102450) to update/disable
tests that depend on hard-closed metrics because of how there may be
changes based on the above described "delayed" (with respect to the
previous implementation) update.