SERVER-100207 Move some resharding files under sharding_coord_d target (#33479)
Move a few resharding files under the `sharding_coord_d` target, and use
`no_undefined_ref_DO_NOT_USE` for it,
so only this subset of the existing `sharding_runtime_d` target needs to
run (at this time) on a large_mem machine.
`no_undefined_ref_DO_NOT_USE` is needed because of symbols defined in
`sharding_runtime_d` which without
that parameter would result in link failure in the dynamic build case.