SERVER-97357 deduplicate transaction logging for prod and test cases (#33244)
Prior to this commit, transaction participant logging had 5 mostly
duplicated code paths that handled "return logs as string" and "return
logs as BSON" separately, etc.

This commit refactors and deduplicates that logic to leave the
production path as efficient as it was originally (logging all
attributes with minimal copying to `logv2::DyanamicAttributes`) and the
test path to produce equivalent BSON objects with a separate but single
implementation.

One implementation for production, one in test to verify the production
code against, all the other parameterizations necessary for test
coverage are now handled by transforming the original log objects as
required.