SERVER-101669 Track interrupt latency in Operation Context (#33229)
Following on from discussions on this
[pr](https://github.com/10gen/mongo/pull/32917), the resolution was that
it makes sense for the OperationContext to track the time taken to
interrupt an operation. This is one of the few roles that is exclusive
to the operation context, so it doesn't make sense to push it into a
decorator.

This moves the existing implementation out of the RecoveryUnit, and
records it as a more general statistic, which still logs to the slow
log. This removes one of the places the WiredTigerUtil.cpp relies on
RecoveryUnit