SERVER-102405 Remove OP_QUERY-targeted optimization in AsyncResultsMerger (#33662)
The `AsyncResultsMerger` can optionally use a `batchSize` for the
`getMore` commands it sends to the shards.

The batchSize is only used when it is set in the `AsyncResultsMerger`.
This only happens when the batchSize is transferred from the
`ClusterClientCursorParams`. The batchSize in the latter is set in a few
cases (see SERVER-102405 ticket comments for details), but it is not set
for some others commands (e.g. change stream cursors).

If the batchSize is set, the `AsyncResultsMerger` will compare it
against the number of documents already fetched from the remote, and
will try to only fetch the remaining ones (i.e. batchSize -
`remote.fetchedCount`). As the value of `remote.fetchedCount` will only
ever increase and exceed the value of `batchSize`, at some point it will
be using just `batchSize`.
We can simplify this code to always use the configured batchSize, if
set.
The original motivation to track the number of fetched documents per
remote and dynamically adjust the batchSize seems to have been to avoid
the OP_QUERY-era SPLIT_SORT full-sort path, which does not exist
anymore.