SERVER-101134 Add backoff to mongos write retries (#33661)
Mongos will limit the number of times it can retry a write without
completing any operations when the reason for lack of progress is
outside of its control. In such cases where the failure is outside of
its control, mongos has no choice but to retry and hope that the problem
was resolved by other nodes by the time the retry attempt is made.
However, mongos does not wait before performing these retries. This can
lead to cases where mongos retries too quickly and exhausts its retry
limit before the other nodes have a chance to unblock progress.

This commit mitigates this by adding an exponential backoff between
retry attempts where the reason for lack of progress is outside of
mongos's control.