SERVER-97171: Handle large change stream event (#32237)
It is possible to read a large change stream event (>16MB) which results
in `BSONObjectTooLarge` error and fails the processor. Change stream
allows the split of large events through `$changeStreamSplitLargeEvent:`
in the source config.
This PR does the following,

- Adds the `$changeStreamSplitLargeEvent:` to $source.config.pipeline
for change stream source by default
- Handles the reading and stitching of the fragments of change stream
event larger than 16MB

`$changeStreamSplitLargeEvent:` is supported starting MongoDB 7.0 (and
6.0.9)

https://www.mongodb.com/docs/manual/reference/operator/aggregation/changeStreamSplitLargeEvent/