SERVER-102544: Properly handle resume token in the middle of a change stream batch (#33726)
Since day 0 we've been improperly dealing with "mid batch" resume tokens
in the ChangeStreamSourceOperator. We were advancing the iterator before
calling get_resume_token, so we were effectively saving the resume token
for the _next_ document that we hadn't read yet. That document would be
skipped on recovery. This can cause correctness issues if the final
checkpoint is written while in the middle of processing a change stream
batch.

We were already dealing with post batch resume tokens correctly, which
mitigates the chance of a customer running into this.