SERVER-99662 Implmement building msi's in bazel (#33677)
Currently msi's only build in scons. Implement building of msi's in
bazel, and upgrade wix version to latest. This does not move CI over to
building msi's through bazel.

This is a fix to a previous PR that got reverted, only significant
change is using vswhere.exe rather than the environment variable
VCINSTALLDIR to find where visual studio is installed.