SERVER-101739 Make ShardRegistry accept inconsistent topologyTime in benign case (#33257)
Due to SERVER-63742, it is possible that an inconsistent topologyTime is
gossiped in the cluster. When config.shards has no entries with
topologyTime, due to coming from a version pre-5.0, it is known that
this inconsistency is benign. This patch makes it accepted until we heal
the cluster during setFCV(9.0). After 9.0 is branched, we can remove the
relaxation.