SERVER-100885: Disable exception capturing for logging in non-plain format (#33722)
Exception capturing works by finding the first and the last line of
exception in the log. In Structured Logging Format the exceptions are
printed as a single JSON. There is very little value in keeping the
summaries of exceptions when one could run the whole log through `jq`
and filter out all the interesting bits.

Co-authored-by: Serhii Lysenko <serhiilysenko@mongodb.com>