SERVER-102113 Disable and fix timeseries tests that rely on not marking soft-closed and archived buckets rollover reasons (#33654)
Set to 0 which is correct with the new behavior and comment it out to
remove redness during integration runs
Also took a look at `timeseries_insert_compression_failure.js` (another
test that looks at `numBucketsClosedDueToReopening`) and it isn't
impacted by these changes.

This change isn't from from the reopening aspect; we now always mark the
`rolloverReason` of the bucket in
[`determineBucketRolloverForMeasurement`](https://github.com/10gen/mongo/blob/master/src/mongo/db/timeseries/bucket_catalog/bucket_catalog.cpp#L178),
[`rollover`](https://github.com/10gen/mongo/blob/master/src/mongo/db/timeseries/bucket_catalog/bucket_catalog_internal.cpp#L1471),
and
[`tryToInsertIntoBucketWithoutRollover`](https://github.com/10gen/mongo/blob/master/src/mongo/db/timeseries/bucket_catalog/bucket_catalog_internal.cpp#L869),
so when we range through [this
loop](https://github.com/10gen/mongo/blob/master/src/mongo/db/timeseries/bucket_catalog/bucket_catalog_internal.cpp#L719)
in `loadBucketintoCatalog`, we don't attempt to re-open the bucket
because it doesn't have a `kNone` rollover reason.

Before, we had a case where sometimes we wouldn't mark the bucket's
rollover reason as soft-closed or archived in
[`insertIntoBucket`](https://github.com/10gen/mongo/blob/master/src/mongo/db/timeseries/bucket_catalog/bucket_catalog_internal.cpp#L786).

This suggests at this point that we will be keeping soft-closed or
archived buckets in memory for longer. I would like to understand if
this is a tradeoff we considered/if we would like this.