SERVER-102537: Temporarily ignore function ubsan checks for ASP third_party dependencies (#33780)
This PR applies `-fno-sanitize=function` to the ASP third-party
dependencies because we're running into function pointer UBSAN problems
on the v5 toolchain with the mongoc driver version we're using
([BF-36532](https://jira.mongodb.org/browse/BF-36532)).

The proper fix is to upgrade the driver, but that will take some time.
This is just a temporary fix. I've added a TODO to reflect this