SERVER-102566 Make db_timeseries_write_ops_internal_test use fixed timestamps for UTC datetime bson type (#33807)
The test failure was due to not using a hard-coded timestamp for the
datetime meta field values. This would enable us to have different meta
fields which would lead to measurements that should have been in the
same batch to be in different batches.
Checked it reproduced by adding a sleep between each measurement we
create in _generateMeasurementWithMetaField for the timestamp case and
saw that this leads to us to make each date bson type measurement in its
own batch.