SERVER-102714 Clear foreign collection plan cache to avoid test failures (#33899)
Follow-up to SERVER-101804.

In the test we have a branch that assumes that we executed in SBE. But
in case of stale router data this test might execute in Classic. In this
case, it will create a plan cache entry in foreign collection for the
Classic's sub-query. This will fail an assertion later that foreign
collection plan cache is entry.

In SERVER-101804 we fixed that by executing a warm-up query to update
router data. But this query still may create a foreign collection entry,
so we clean it after the warm-up query.