SERVER-102715 [SBE] Don't release storage engine resources if saveState()/restoreState() throws (#33901)
SERVER-101591 changed saveState() and restoreState() to release storage
engine resources when an exception was thrown, but this caused problems
for cases where a WriteConflict exception is thrown and caught during
yielding and execution continues.

This PR backs out the problematic part of SERVER-101591, while keeping
the changes to the CollectionRef class that were needed to address the
immediate problem from BF-36692 and BF-37044.