SERVER-81208 Use checkpoint cursor timestamp when opening backup cursor (#31485)
This PR removes old checkpoint detection code when opening up a backup
cursor in favour of using the newly provided query_timestamp API to
directly query the stable checkpoint timestamp to open the backup cursor
with. This enables us to avoid having to re-check the checkpoint
timestamp validity.

The semantics of this requires us to pin the oldest timestamp and take a
checkpoint while opening the backup cursor to ensure that the oldest
timestamp doesn't move beyond the read timestamp while we are opening
the cursor on the catalog. This also involves updating and fixing some
tests that didn't have this prior assumption.