SERVER-102210 Add code comment for passing a nullptr (#33933)
`_solution` is allowed to be a nullptr when passing it to
`plan_explainer_factory::make(...)`. Verified by experiment (adding
`invariant(_solution);` to the code) that we invoke
`plan_explainer_factory::make(...)` with `_solution` being a nullptr in
plenty of tests.
Added a code comment and workaround stating that a nullptr `_solution`
is a valid here. This is needed to silence a false positive in Coverity.