SERVER-101615 SERVER-102504 Drain ongoing coordinators after transitioning to kUpgrading and clone DB metadata (#33301)
After SERVER-101537, this will use a more precise mechanism than just
`waitForOngoingCoordinatorsToFinish()`.

This PR also adds `enable_on_transitional_fcv: true` to
`featureFlagCreateDatabaseDDLCoordinator` and extends the
`FixedFCVRegion` over the complete duration of the legacy createDatabase
command, to serialize with it. The reason we need to do this is to also
serialize calls to the non-coordinator version of createDatabase, i.e.
make sure that there are no such commands running during `kUpgrading`,
since authoritative databases relies on the coordinator implementation.
For this same reason, we need the coordinator during `kUpgrading`, hence
`enable_on_transitional_fcv: true`.