SERVER-100723 Configure scheduled test builders on m8g/AL2023 for MongoDB master (#33378)
Newly created variants on v6.0 to test correctness on Graviton 4.

Ticket: https://jira.mongodb.org/browse/SERVER-100723

This newly added variants will run nightly. On master, this will be
<10.0% of our testing schedule for Amazon Linux variants. The testing
schedule calculation is `(# of times new variants are scheduled in a
week)/(# of times any AL variant is scheduled in a week)`, which may be
correlated with budget considering the variants all run somewhat similar
tasks.

AL2023 is used as the OS, because it's the OS we expect to dominate
Atlas by the time of our ARM v9 migration.