SERVER-82619 refactor inReplicationRecovery to use RAII type (#33777)
Previously inReplicationRecovery was a boolean flag on ServiceContext,
but is now a uint64_t. This supports recursive setting of this flag as a
reference count, with zero indicating that nothing is in progress
performing replication recovery.

This change was done to clean up multiple recursive settings and
invariant checks to the flag to ensure it is set only in the proper
contexts. This new implementation is easier to read and confirm
correctness by inspection in addition to test coverage.