SERVER-95229: Support non-removable window implementations for $minMaxScaler window function (#32355)
Adds full support for all cases of `$minMaxScaler` window function for
`$setWindowFields` agg stage (both removable and non-removable window
implementations). Removable implementations we're already supported in
the [previous PR](https://github.com/10gen/mongo/pull/27507).

The bulk of this PR is adding in custom `WindowFunctionExec` classes for
non-removable windows, for both document and range based bounds (that is
`WindowFunctionExecMinMaxScalerNonRemovable` and
`WindowFunctionExecMinMaxScalerNonRemovableRange`), and then installing
them in `translateMinMaxScaler` (in `window_function_exec.cpp`).

To accomplish this, we refactor the existing
`WindowFunctionExecNonRemovable` / `WindowFunctionExecNonRemovableRange`
classes into `WindowFunctionExecNonRemovableCommon` /
`WindowFunctionExecNonRemovableRangeCommon` classes, which held the
majority of the logic (iterating over the window bounds) we needed to
reuse. These common classes now define virtual interfaces that describe
how to update, reset, and get the current value of the window. The
existing `WindowFunctionExecNonRemovable` /
`WindowFunctionExecNonRemovableRange` classes are re-implemented in
terms of these common classes, and the new
`WindowFunctionExecMinMaxScalerNonRemovable` /
`WindowFunctionExecMinMaxScalerNonRemovable` also are implemented in
terms of the common classes as well, with the minimum set of logic
specific to computing `$minMaxScaler`.

Some other refactoring: Adds in a
`window_function_min_max_scaler_util.h`/`cpp` file for common structures
/ functions that apply to both removable and non-removable
implementations.

**Note:** some additional TODO tickets, that can be done in parallel /
independently, are introduced in this PR to reduce the scope, merge more
quickly, and unblock other tickets that depend on this:
- [SERVER-102248](https://jira.mongodb.org/browse/SERVER-102248): Adding
in memory tracker to `WindowFunctionExecMinMaxScalerNonRemovable` /
`WindowFunctionExecMinMaxScalerNonRemovable` (planning on consulting
with @wolffcm, as I'm not currently familiar with memory tracking).
- [SERVER-102701](https://jira.mongodb.org/browse/SERVER-102701): Unit
testing `WindowFunctionExecMinMaxScalerNonRemovable` /
`WindowFunctionExecMinMaxScalerNonRemovable` (these are already
throughly tested indirectly through end-to-end js tests in
`min_max_scaler_output_values.js`)
- [SERVER-102703](https://jira.mongodb.org/browse/SERVER-102703):
Investigate the behavior of reversing the sort order. Currently we
assume the sort order for $minMaxScaler is ascending (1). We need to
make sure this new window function's behavior is consistent with others
window functions (it should either fail correctly or run correctly).
Note that this PR does not open this question, or alter this behavior
(it existed in the previous removable window implementation PR), so it
makes sense to open a separate ticket for investigation, instead of
further complexifying an already large change.

All of this logic is feature flag protected under
`featureFlagSearchHybridScoringFull` and all of these TODOs will be
completed before `$minMaxScaler` / `$scoreFusion` is released in 8.2.