SERVER-102768 Add debug output to sbe_explain_rejected_plans.js test (#33974)
[BF-37073](https://jira.mongodb.org/browse/BF-37073) contained a test
failure in the `sbe_explain_rejected_plans.js` test.

The cause of the test failure is currently unclear.
Unfortunately the failing assertion in the test is not very helpful, as
it only compares a non-zero array length to a zero array length, but
does not provide any details about the actual explain outputs.
This PR adjusts the assertions in the test so that they provide more
output should the assertions be violated.