SERVER-100913 Connect new timeseries batched write path (#33416)
This new architecture of staging timeseries writes in batches instead of
one-at-a-time should yield benefits to user-observable performance and
debugging. Any performance benefits will be targeted at large batches of
inserts (i.e. `insertMany` instead of `insertOne`) with the same, or a
relatively small number of, metafield(s).

This commit will also introduce a slight natural ordering change because
documents within a user's batch will be sorted on time before being
inserted into the buckets collection which has benefits to bucketing and
therefore performance and compression.