SERVER-100617 removed CoerceToBool from optimization of $and expressions to correctly evaluate conditional in $expr (#33021)
We only use `$coerceToBool` when we have `$and` or `$or` with two
children - a non-constant expression and a constant True, e.g.
`{$expr:{$and:[ {$eq:["$_id",15]}, true ]}}`. In such cases, the
constant gets removed and the remaining expression gets converted to
`$coerceToBool`, e.g. `{$expr:{$coerceToBool: {$eq:["$_id",15]} }}`.
This is equivalent to `$and` or `$or` with a single (non-constant)
child, e.g. `{$expr:{$and:[ {$eq:["$_id",15]}]}}`. Since `$coerceToBool`
is not used in any other scenarios, changes in this PR remove it
completely.

---------

Co-authored-by: Hana Pearlman <hana.pearlman@mongodb.com>