SERVER-101468 Update certificate generation to be cross-platform and deterministic (#33814)
Adds a new directory, x509/, to the root. This directory contains:

- mkcert.py: A script which deterministically generates certificates,
CRLs, and digests based on a definitions file
- Based on jstests/ssl/x509/mkcert.py, but ported to use the python
cryptography library for better portability and support for
deterministic generation
- certs.yml: A definitions file which defines all of the certificates
needed for our testing
- Almost jstests/ssl/x509/certs.yml, but with some changes to fit the
new certificate definition format
- apple-certs.yml: A definitions file which defines a certificate chain
that needs to be generated and added to the trusted keychain during
MacOS provisioning.
- static/: A directory currently only containing key files that are used
by mkcert.py during generation. Later, it will contain all of the
certificates, digests, CRLs, etc. which are not generated by mkcert.py.

Adds a no-passthrough test, mkcert_is_deterministic.js, which ensures
that running certificate generation twice produces the same results both
times, and also that pkcs12 generation works.

Updates the cryptography and pyOpenSSL python packages, as the new
versions are required.

Removes special handling for the cryptography and pyOpenSSL packages on
IBM Z and Power machines as well as the "oldcrypt" poetry flag; these
are no longer necessary as the underlying issue was fixed.