SERVER-102927 Skip pdb copy step in unit test gather script if binary is not an exe (#34144)
It looks like the logic for archiving PDB symbol files causes an error
on Linux platforms and prevents the .debug symbol files from being
archived correctly. This prevents the C++ stacktraces in the failing
logs from being symbolized and also prevents the core dump from being
utilized.

```
[2025/03/21 21:47:54.195] Install file: "bazel-out/aarch64-dbg/bin/src/mongo/db/modules/enterprise/src/streams/exec/tests/s3_emit_operator_test" as "build/install/bin/s3_emit_operator_test"
[2025/03/21 21:48:01.178] Install file: "bazel-out/aarch64-dbg/bin/src/mongo/db/modules/enterprise/src/streams/exec/tests/s3_emit_operator_test.debug" as "build/install/bin/s3_emit_operator_test.debug"
...
[2025/03/21 21:50:43.101] Files to save: build/install/bin/s3_emit_operator_test
[2025/03/21 21:50:43.101] Direct Copy build/install/bin/s3_emit_operator_test to dist-unittests/bin/s3_emit_operator_test
[2025/03/21 21:50:43.107] PDB Copy build/install/bin/s3_emit_operator_test to dist-unittests/bin/s3_emit_operator_test
[2025/03/21 21:50:43.108] cp: cannot create regular file 'dist-unittests/bin/s3_emit_operator_test': Permission denied
```

Skip this process entirely if the file doesn't end in .exe since that's
what the sed depends on to generate the pdb path.