Add stage descriptors to the extension API; implement SDK (#34137)
This creates a new stage structure that is incompletely decoupled from
`DocumentSource`. Each stage has 3 components:
* A descriptor which defines the name and pipeline-independent
properties.
* A bound descriptor which adds stage definition and pipeline context to
the stage descriptor and is created from a descriptor.
* An executor that is used to process documents and is created from a
bound descriptor.

This PR contains the header, rust SDK changes, and a trivial example
(`$addSomeCrabs`) but is missing some things in
an attempt to control the size of the PR:
* There is no extension host implementation.
* The SDK executor is the existing `AggregationStage` implementation,
which doesn't align with the long-term vision.
* Transform stages do not yet accept a source as part of executor
creation.
* I haven't reimplemented any other stages.