SERVER-102938 Separate the block and scalar mode flows in SBE $group stage builder (#34153)
Refactor the SBE $group stage builder code to separate the block
processing (BP) from the scalar (aka "normal") flow. The resulting
separation will increase understandability and relax the restriction
that block and scalar data structures must be essentially the same where
they overlap. It will also eliminate quite a bit of code, because it
will no longer need to deal with "backing out" partial progress of block
stage building, which can be abandoned at several different points when
something BP does not support is encountered, not all of which can be
checked for up front.