SERVER-96986 Add timeout to benchmark and fix mac binaries being regular tar files (#34213)
Mac binaries were not being gzipped, they were just being created as
regular tar files which was causing signing to fail and packages to be
much larger. Fix this by passing the extension to tar rule.

Also add timeout on ! AL2023.3 arm64 Enterprise (SEP Benchmark) as it is
part of the commit queue and can hit the infinite building bug.