SERVER-102718 Re-enable all test cases in jstests/auth/lib/commands_lib.js (#34105)
SERVER-95901 mistakenly disabled all by the last test case in
`commands_lib.js`. Since then, several tests have been added or modified
and began silently failing. This PR re-enables all the test cases and
fixes the broken ones.

This PR partially reverts SERVER-97725, removing the `searchCoordinator`
role from the set of builtin roles that this library tests. This was
done because fixing that properly will require modifying many more tests
in this file (each test case which manually specifies the set of roles,
of which there are >300, may need to be updated). This is best done in a
follow-up so we can re-enable the tests sooner.