SERVER-102979 Disable tests from running in multiversion suites (#34188)
[SERVER-100260](https://jira.mongodb.org/browse/SERVER-100260) (8.2)
added new functionality to Queryable Encryption $lookup that breaks the
existing `aggregate_lookupMultiSchema.js` test when run in a 8.1+8.2
multiversion cluster. That test is tagged with requires_fcv_81 and needs
to be updated to requires_fcv_82. Similarly, `findandmodify_wce.js` was
added in 8.2, but has a requires_fcv_81 tag, so it is now updated to
requires_fcv_82.
Lastly, `fts_spilling.js` is not suited to run in a multi sharded
cluster fixture as
sharded_collections_query_shape_hash_stability_multiversion does, so it
is excluded from the suite.