SERVER-86688 Emit additional change event fields, regardless of showExpandedEvents (#33843)
This change will expose the following extra fields in change stream
change events by default, which were previously only expanded when
showExpandedEvents was set. Starting with version 8.2, these fields will
be emitted unconditionally, regardless of the value of
showExpandedEvents:

* `collectionUUID` for DML events (`insert`, `update`, `replace`,
`delete`)
* `commitTimestamp` for DML events that are part of a prepared /
distributed transaction
* `updateDescription.disambiguatedPaths` field for update events
 
The following fields will be shown if emitting the respective events is
also enabled:

* `collectionUUID` for DDL and control events (except `invalidate`,
`dropDatabase`, `endOfTransaction`)
* `operationDescription` for DDL and control events (`create`, `rename`,
`createIndexes`, `startIndexBuild`, `abortIndexBuild`, `dropIndexes`,
`modify`)
* `nsType` for DDL create events

There will be no way to disable emitting the extra fields in change
stream events in 8.2 and higher.
If the fields are not needed by consumers, they can be projected out of
the events using a `$project` stage with exclusions.

Note that we still have to keep the `showExpandedEvents` flag for change
streams even in the future, as the flag will still control if DDL events
will be emitted or not. This will not be changed by this PR.

We also need to execute several tests with the `showExpandedEvents` flag
set to ensure that test results are identical between 8.2 (which emits
additional fields unconditionally) and older releases (which only emit
additional fields if `showExpandedEvents` is used) in multiversion
tests.

---------

Co-authored-by: Xuerui Fa <xuerui.fa@mongodb.com>