SERVER-102989 Refactor PBT runner to use a workload model (#34208)
Instead of passing two models to the PBT runner, a collection model and
an aggregation model, we pass one "workload" model consisting of both.

This allows PBTs to be more flexible, since the collection model and agg
model don't have to be independent. Example is provided in the readme.

Most of this PR is just wrapping the existing collection models and agg
models into one object.