Extension host code for plugin descriptors (#34273)
Introduce `DocumentSourceExtension` as a replacement for
`DocumentSourcePlugin`.

This doc source registers descriptors and uses the descriptor API to
create the necessary objects. This allows us to
extend the so that extension stages can use stage constraints. At the
moment all objects (including executor) are
created before creating the document source but this can be changed in
the future.

Change $echoOxide and $pluginSort to use the descriptor API and set
stage constraints appropriately.