SERVER-100334 Fix StackLocator functionality broken by v5 toolchain ASAN (#34132)
In the v5 toolchain, clang's ASAN enables detect_stack_use_after_return
by default, which wasn't the case for v4. This has the effect of making
automatic storage duration variables not necessarily be allocated on the
stack. The StackLocator object works by inspecting the address of such a
variable and measuring the distance from there to the end of the stack
space. Since detect_stack_use_after_return makes it so such variables
aren't on the stack, this doesn't work.

We initially just disabled detect_stack_use_after_return. This change
re-enables it and modifies `StackLocator` to use
`__builtin_frame_address(0)` (where available) to find the frame
pointer. This doesn't give the exact same results it gave before - the
old version would use the address of a particular variable on the stack,
and this one uses the address of a particular frame. For obvious
reasons, these may not line up exactly.

The sole user of `StackLocator` is the javascript engine, which uses it
to set an upper limit on the stack space that the engine will use. It
already uses a safety margin of at least 64kB, so the difference between
using the frame pointer and using the address of a variable on the stack
shouldn't be enough to cause any issues.