SERVER-100421 Fix resharding logging for atlas ingestion (#34129)
This commit makes a number of changes:

- Fix the resharding complete log used by atlas log ingestion to report
values even when resharding fails by ensuring that the recipient updates
the coordinator of its metrics even when aborted and ensuring that the
coordinator reports based on the latest version of its on-disk state
document.
- Add the failure reason (if resharding failed) and the user supplied
UUID (if one was specified) to the completion log.
- Fix the completion log's reporting of total documents and total bytes
to be based on the approximate total number of documents and bytes to
copy according to the donors instead of the actual amounts copied by
recipients.
- Report recipient metrics under well known field names instead of their
shard ids.
- Report the time each recipient spent in each of its phases as part of
recipient metrics, including the time it spent building indexes.
- Stop reporting time spent building indexes as part of the coordinator
metrics since the coordinator does not build indexes; previously this
value was always reported as 0.
- Replace the existing sharding test for completion metrics with a
core_sharding test.
- Improve the createChunks test helper to produce chunks that more
evenly distribute the provided range.