SERVER-99915 FLE2 findAndModify should also return WCE from transaction API result (#34001)
This reverts the changes in bad1a746f41fb17fc5c0fe867f00fe74a1f6ac52,
and provides an alternative way of handling write concern errors for the
mongos case when a command error occurs. This new fix builds on the
fixes made in SERVER-102024 (for the mongod case) and SERVER-97470 (for
correcting the WCE reporting in mongos for unencrypted findAndModify).