SERVER-101491 Quiesce in JWKFetcherImpl when remote returns 5xx error (#33348)
Currently the server will limit a client who spams the server trying to
authenticate with an unknown KID to mitigate the risk of getting
firewall-banned from the IDP. The time period in which a request must
pass is set by the server parameter `JWKSMinimumQuiescePeriodSecs`.

Still, there are several cases in which a server could get
firewall-banned due to not being affected by the current implementation
of the quiesce period:

1. An authenticated client calling `oidcRefreshKeys` in a loop would
ignore the quiesce period and fetch the IDP on each command.
2. `JWKRefreshJob` failing to fetch an IDP would go on a loop until
successful. If the IDP changed its API route or if the server is down,
we could still get banned by their cloud provider if we fetch the IDP in
a loop until successful.

The changes in this PR fix these two scenarios.

The original ticket suggested only throttling the fetches if the IDP was
returning a 5xx error, but I came up with these other scenarios during
testing. The only risk I see is a client application with
`JWKSMinimumQuiescePeriodSecs` could break if they expect that running
`oidcRefreshKeys` will always succeed.