SERVER-102100 Correct numBucketsOpenedDueToMetadata (#34343)
Added flipping numBucketsOpenedDueToMetadata to false in:
- `findAndRolloverOpenBuckets` if we have at least one bucket in the
open bucket key set
- `getReopenedBucket` to account for the bucket-reopening case; this was
a bit more nuanced then what we originally thought. I saw that when I
put the check in potentiallyReopenBucket, sometimes we would also say
there was no metadata even in the case that we didn't get an document
from the [reopening candidate
](https://github.com/10gen/mongo/compare/master...stephanie.eristoff/SERVER-102100#diff-e6ace70884a6b4495edc2dc5e9fdcdd986da02f56dcfdefbd1c3f35f8ab68f95R709).

Notes on the test changes:
- timeseries_insert_compression_failure.js changes are because we have
the case that for the malformed bucket, we do have a bucket that has the
matching metadata. Before, we didn't count the malformed bucket as
having the same metadata.
- timeseries_collstats.js is adding back the testing (only
`numBucketsOpenedDueToMetadata`) 🥳
- unit test changes to account for new arg in
`findAndRolloverOpenBuckets`